Make 2 new functions in js.
A new function called getComputerChoice was added in js. The function
works by creating a new variable  i who gets assigned a random integer
between 0 and 2 inclusively. Each of these integers is assigned to either
rock, paper, or scissors each time the function is called. The
getComputerChoice is assigned to the computerSelection. Now to get the
player selection I set the variable playerSelection to a prompt and since
we will be working in all lowercase i make sure anything input into that
prompt turns to lowercase using .toLowerCase. Now A new function is created
called playRound. The function has 2 parameters, the computerSelection
and the playerSelection. Using if and else if statements we go through
all options of rock,paper,scissors from the player vs the computer and
return either "YOU WIN!!" , "YOU LOSE!!" or "IT'S A TIE".