Add a new function and rework previous ones
Four new variable were created to work into the playRound function. The
new variables include, playerWins, computerWins, tie and result. The new
variables are meant to used for the return to each combination in the
playRound function. Rather than use actual strings as the return, the
variables will have the strings with the according win/lose message. The
result variable will hold the playerWins or computerWins or tie depending
on the situation. Changing the result values this way will help with the
game() function we have to later create. Next up the game function was
created. This function is meant so 5 rounds of rps are played and
will keep count of every the winner every round and announce the loser
at the end of the five rounds. Two new variables playerScore and
computerScore were created to keep score of the game each round. We use
a for loop to loop to 5 rounds of the game. In the for loop we include
the let playerSelection= prompt...; so each round the player is newly
asked to choose. We also include the
let computerSelection = getComputerChoice(); here because if included
prior to here the computerSelection would remain the same each round.
Using if statement I incrased the score of the computer or the player
based according to the result. If either the computerScore or the
playerScore was >=3 that player wins, and the according string would be
printed as the return. To keep score each round console.log() statements
with the computerScore and playerScore were included. Finally in the case
of a tie in the game and if statement outside of the forloop was included
so if the computerScore==playerScore to print a tie, this was done outide
of the for loop because if done inside any time the computerScore is the
same as the playerScore it would end the game and print a tie to the
console.